pyproject.toml
[project]
name = "car-price-predictor"
version = "0.0.1"
description = "This project contains a web app that will predict the price of the cars."
authors = [
    {name = "Huraira"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "missingno (>=0.5.2,<0.6.0)",
    "flask (>=3.1.0,<4.0.0)",
    "pymongo (>=4.12.1,<5.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
